pyproject.toml
[tool.poetry]
name = "mqtt-receiver"
version = "1.0.0"
description = "To test receive message from HiveMQ MQTT"
authors = ["Sharizan Redzuan <sharizan_81@yahoo.com>"]
readme = "README.md"

[tool.poetry.scripts]
dev = "mqtt_receiver.__main__:main"

[tool.poetry.dependencies]
python = "^3.12"
paho-mqtt = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
